+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix typo that prevented using `jsonify`.
+
 ## [2.6.0] - 2024-08-26
 
 ### Added